[feat] split routes into different files
